+++ contracts/babylon/src/msg/contract.rs
@@ -1,8 +1,6 @@
+use babylon_apis::finality_api::Evidence;
 use cosmwasm_schema::{cw_serde, QueryResponses};
 use cosmwasm_std::{Binary, StdError, StdResult, Uint128};
-
-use babylon_apis::finality_api::Evidence;
-
 #[cfg(not(target_arch = "wasm32"))]
 use {
     crate::msg::consumer_header::{ConsumerHeaderResponse, ConsumerHeightResponse},
@@ -10,26 +8,11 @@ use {
     crate::state::config::Config,
 };
 
-const BABYLON_TAG_BYTES: usize = 4;
-
-// Common functions for contract messages
-pub trait ContractMsg {
-    fn validate(&self) -> StdResult<()>;
-    fn babylon_tag_to_bytes(&self) -> StdResult<Vec<u8>>;
-}
-
 #[cw_serde]
 pub struct InstantiateMsg {
     pub network: btc_light_client::BitcoinNetwork,
-    /// A string encoding four bytes used for identification / tagging of the Babylon zone.
-    /// NOTE: this is a hex string, not raw bytes
-    pub babylon_tag: String,
     pub btc_confirmation_depth: u32,
     pub checkpoint_finalization_timeout: u32,
-    /// Whether to send Cosmos zone messages notifying BTC-finalised headers.
-    /// NOTE: If set to true, then the Cosmos zone needs to integrate the corresponding message handler
-    /// as well
-    pub notify_cosmos_zone: bool,
     /// If set, this will instantiate a BTC light client contract
     pub btc_light_client_code_id: Option<u64>,
     /// If set, this will define the instantiation message for the BTC light client contract.
@@ -70,10 +53,8 @@ impl InstantiateMsg {
     pub fn new_test() -> Self {
         Self {
             network: btc_light_client::BitcoinNetwork::Regtest,
-            babylon_tag: "01020304".to_string(),
             btc_confirmation_depth: 1,
             checkpoint_finalization_timeout: 1,
-            notify_cosmos_zone: false,
             btc_light_client_code_id: None,
             btc_light_client_msg: None,
             btc_staking_code_id: None,
@@ -90,16 +71,8 @@ impl InstantiateMsg {
     }
 }
 
-impl ContractMsg for InstantiateMsg {
-    fn validate(&self) -> StdResult<()> {
-        if self.babylon_tag.len() != BABYLON_TAG_BYTES * 2 {
-            return Err(StdError::invalid_data_size(
-                BABYLON_TAG_BYTES * 2,
-                self.babylon_tag.len(),
-            ));
-        }
-        let _ = self.babylon_tag_to_bytes()?;
-
+impl InstantiateMsg {
+    pub fn validate(&self) -> StdResult<()> {
         if self.btc_staking_code_id.is_some() {
             if let (Some(consumer_name), Some(consumer_description)) =
                 (&self.consumer_name, &self.consumer_description)
@@ -125,15 +98,6 @@ impl ContractMsg for InstantiateMsg {
         }
 
         Ok(())
-    }
-
-    fn babylon_tag_to_bytes(&self) -> StdResult<Vec<u8>> {
-        hex::decode(&self.babylon_tag).map_err(|_| {
-            StdError::generic_err(format!(
-                "babylon_tag is not a valid hex string: {}",
-                self.babylon_tag
-            ))
-        })
     }
 }